Fix the get entities in systems. Added support for textures with albedo, normal, specular maps. Reduced warnings.

+++ R3D/src/R3D/Core/Window.cpp
@@ -106,7 +106,7 @@ namespace r3d
 		// TODO
 	}
 
-	void Window::setViewPort(float width, float height) const
+	void Window::setViewPort(double width, double height) const
 	{
 		glViewport(0, 0, (GLsizei) width, (GLsizei) height);
 	}
